Only cleaning up bins if the directory exists

+++ build/clean.sh
@@ -24,11 +24,13 @@ cd ../src/Cyberena
 clean_cmake
 cd - &>/dev/null
 
-if [[ "`ls ../bin | grep -v "Test"`" != "" ]]; then
-	ls ../bin/* 2>/dev/null | grep -v "Test" | xargs rm
+if [ -d ../bin ]; then
+	if [[ "`ls ../bin | grep -v "Test"`" != "" ]]; then
+		ls ../bin/* 2>/dev/null | grep -v "Test" | xargs rm
+	fi
+	if [[ "`ls ../bin/Test`" != "" ]]; then
+		rm ../bin/Test/*
+	fi
+	echo "Cleaned out bin!"
 fi
-if [[ "`ls ../bin/Test`" != "" ]]; then
-	rm ../bin/Test/*
-fi
-echo "Cleaned out bin!"
 exit 0
